+++ src/order.js
@@ -1,5 +1,16 @@
 // @ts-check
 
+import {
+	all_sizes,
+	border,
+	content_items_self,
+	inline_block,
+	min_max,
+	start_end,
+	top_right_bottom_left,
+	x_y_inline_block,
+} from './util'
+
 /**
  * @returns {string[]} Array of property names
  */
@@ -116,7 +127,6 @@ function layout() {
 		'columns',
 		'column-count',
 		'column-fill',
-		'column-gap',
 		'column-rule',
 		'column-rule-color',
 		'column-rule-style',
@@ -220,7 +230,6 @@ function typography() {
 		'text-emphasis-position',
 		'text-emphasis-style',
 		'text-orientation',
-		'text-overflow',
 
 		// white space & word wrapping
 		'white-space',
@@ -401,146 +410,4 @@ function misc() {
 		'page-break-after',
 		'print-color-adjust',
 	]
-}
-
-/**
- * Returns an array of property names with the following suffixes:
- * - top
- * - right
- * - bottom
- * - left
- * @param {string | undefined} [name] Base property name
- * @returns {string[]} Array of property names
- */
-function top_right_bottom_left(name) {
-	const prefix = name !== undefined ? `${name}-` : ''
-	const properties = ['top', 'right', 'bottom', 'left'].map(after(prefix))
-
-	return [name ?? [], properties].flat()
-}
-
-/**
- * Returns an array of property names with the following suffixes:
- * - x
- * - y
- * @param {string} name Base property name
- * @returns {string[]} Array of property names
- */
-function x_y(name) {
-	return ['-x', '-y'].map(after(name))
-}
-
-/**
- * Returns an array of property names with the following suffixes:
- * - inline
- * - block
- * @param {string} name Base property name
- * @returns {string[]} Array of property names
- */
-function inline_block(name) {
-	return ['-inline', '-block'].map(after(name))
-}
-
-/**
- * Returns an array of property names with the following suffixes:
- * - x
- * - y
- * - inline
- * - block
- * @param {string} name Base property name
- * @returns {string[]} Array of property names
- */
-function x_y_inline_block(name) {
-	return [name, ...x_y(name), ...inline_block(name)]
-}
-
-/**
- * Returns an array of property names with the following suffixes:
- * - inline
- * - inline-start
- * - inline-end
- * - block
- * - block-start
- * - block-end
- * @param {string} name Base property name
- * @returns {string[]} Array of property names
- */
-function inline_block_with_start_end(name) {
-	return inline_block(name).flatMap(start_end)
-}
-
-/**
- * Returns an array of all sizes for the given property name.
- * @param {string} name Base property name
- * @returns {string[]} Array of property names
- */
-function all_sizes(name) {
-	return [...top_right_bottom_left(name), ...inline_block_with_start_end(name)]
-}
-
-/**
- * Returns an array of property names with the following suffixes:
- * - start
- * - end
- * @param {string} name Base property name
- * @returns {string[]} Array of property names
- */
-function start_end(name) {
-	return ['', '-start', '-end'].map(after(name))
-}
-
-/**
- * Returns an array of property names with the following prefixes:
- * - min
- * - max
- * @param {string} name Base property name
- * @returns {string[]} Array of property names
- */
-function min_max(name) {
-	return ['', 'min-', 'max-'].map(before(name))
-}
-
-/**
- * Returns an array of property names with the following formats:
- * - border-{name}
- * - border-{name}-width
- * - border-{name}-style
- * - border-{name}-color
- * @param {string | undefined} [name] Base property name
- * @returns {string[]} Array of property names
- */
-function border(name) {
-	const infix = name !== undefined ? `-${name}` : ''
-	return ['', '-width', '-style', '-color'].map(after(infix)).map(after('border'))
-}
-
-/**
- * Returns an array of property names with the following suffixes:
- * - content
- * - items
- * - self
- * @param {string} name Base property name
- * @returns {string[]} Array of property names
- */
-function content_items_self(name) {
-	return ['-content', '-items', '-self'].map(after(name))
-}
-
-/**
- * Returns a function that concatenates the given prefix with the provided name.
- *
- * @param {string} name - The name to concatenate with the prefix.
- * @returns {(suffix: string) => string} A function that takes a prefix and returns the concatenated string.
- */
-function before(name) {
-	return (prefix) => `${prefix}${name}`
-}
-
-/**
- * Returns a function that returns a string with the given suffix.
- * @param {string} name - The name to concatenate with the suffix.
- * @returns {(suffix: string) => string} A function that takes a suffix and returns the concatenated string.
- */
-function after(name) {
-	return (suffix) => `${name}${suffix}`
 }
